+++ src/main/java/com/gxdxx/instagram/config/s3/S3Uploader.java
@@ -3,7 +3,7 @@ package com.gxdxx.instagram.config.s3;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.gxdxx.instagram.exception.FileProcessingException;
+import com.gxdxx.instagram.global.error.FileProcessingException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
