+++ src/main/java/com/gxdxx/instagram/domain/chatroom/api/ChatRoomController.java
@@ -3,7 +3,7 @@ package com.gxdxx.instagram.domain.chatroom.api;
 import com.gxdxx.instagram.dto.request.ChatRoomListRequest;
 import com.gxdxx.instagram.dto.response.ChatRoomResponse;
 import com.gxdxx.instagram.dto.response.ErrorResponse;
-import com.gxdxx.instagram.exception.InvalidRequestException;
+import com.gxdxx.instagram.global.error.InvalidRequestException;
 import com.gxdxx.instagram.domain.chatroom.application.ChatRoomQueryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
