+++ src/main/java/com/gxdxx/instagram/domain/chatroom/exception/ChatRoomNotFoundException.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.exception;
+package com.gxdxx.instagram.domain.chatroom.exception;
 
 public class ChatRoomNotFoundException extends RuntimeException {
 }
