+++ src/main/java/com/gxdxx/instagram/domain/comment/application/CommentCreateService.java
@@ -5,8 +5,8 @@ import com.gxdxx.instagram.dto.response.CommentRegisterResponse;
 import com.gxdxx.instagram.domain.comment.domain.Comment;
 import com.gxdxx.instagram.domain.post.domain.Post;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.exception.PostNotFoundException;
-import com.gxdxx.instagram.exception.UserNotFoundException;
+import com.gxdxx.instagram.domain.post.exception.PostNotFoundException;
+import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.comment.dao.CommentRepository;
 import com.gxdxx.instagram.domain.post.dao.PostRepository;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
