+++ src/main/java/com/gxdxx/instagram/domain/comment/exception/CommentNotFoundException.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.exception;
+package com.gxdxx.instagram.domain.comment.exception;
 
 public class CommentNotFoundException extends RuntimeException {
 }
