+++ src/main/java/com/gxdxx/instagram/domain/follow/application/FollowDeleteService.java
@@ -3,8 +3,8 @@ package com.gxdxx.instagram.domain.follow.application;
 import com.gxdxx.instagram.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.follow.domain.Follow;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.exception.FollowNotFountException;
-import com.gxdxx.instagram.exception.UserNotFoundException;
+import com.gxdxx.instagram.domain.follow.exception.FollowNotFountException;
+import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.follow.dao.FollowRepository;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import lombok.RequiredArgsConstructor;
