+++ src/main/java/com/gxdxx/instagram/domain/follow/exception/FollowNotFountException.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.exception;
+package com.gxdxx.instagram.domain.follow.exception;
 
 public class FollowNotFountException extends RuntimeException {
 }
