+++ src/main/java/com/gxdxx/instagram/domain/message/application/MessageCreateService.java
@@ -5,8 +5,8 @@ import com.gxdxx.instagram.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.chatroom.domain.ChatRoom;
 import com.gxdxx.instagram.domain.message.domain.Message;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.exception.InvalidRequestException;
-import com.gxdxx.instagram.exception.UserNotFoundException;
+import com.gxdxx.instagram.global.error.InvalidRequestException;
+import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.chatroom.dao.ChatRoomRepository;
 import com.gxdxx.instagram.domain.message.dao.MessageRepository;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
