+++ src/main/java/com/gxdxx/instagram/domain/message/application/MessageQueryService.java
@@ -5,9 +5,9 @@ import com.gxdxx.instagram.dto.response.MessageListResponse;
 import com.gxdxx.instagram.dto.response.MessageResponse;
 import com.gxdxx.instagram.domain.chatroom.domain.ChatRoom;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.exception.ChatRoomNotFoundException;
-import com.gxdxx.instagram.exception.UnauthorizedAccessException;
-import com.gxdxx.instagram.exception.UserNotFoundException;
+import com.gxdxx.instagram.domain.chatroom.exception.ChatRoomNotFoundException;
+import com.gxdxx.instagram.global.auth.UnauthorizedAccessException;
+import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.chatroom.dao.ChatRoomRepository;
 import com.gxdxx.instagram.domain.message.dao.MessageRepository;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
