+++ src/main/java/com/gxdxx/instagram/domain/post/application/PostFeedQueryService.java
@@ -4,7 +4,7 @@ import com.gxdxx.instagram.dto.request.PostFeedRequest;
 import com.gxdxx.instagram.dto.response.FeedResponse;
 import com.gxdxx.instagram.dto.response.PostFeedResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.exception.UserNotFoundException;
+import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.post.dao.PostRepository;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import lombok.RequiredArgsConstructor;
