+++ src/main/java/com/gxdxx/instagram/domain/post/application/PostUpdateService.java
@@ -3,8 +3,8 @@ package com.gxdxx.instagram.domain.post.application;
 import com.gxdxx.instagram.dto.request.PostUpdateRequest;
 import com.gxdxx.instagram.dto.response.PostUpdateResponse;
 import com.gxdxx.instagram.domain.post.domain.Post;
-import com.gxdxx.instagram.exception.PostNotFoundException;
-import com.gxdxx.instagram.exception.UnauthorizedAccessException;
+import com.gxdxx.instagram.domain.post.exception.PostNotFoundException;
+import com.gxdxx.instagram.global.auth.UnauthorizedAccessException;
 import com.gxdxx.instagram.domain.post.dao.PostRepository;
 import com.gxdxx.instagram.config.s3.S3Uploader;
 import lombok.RequiredArgsConstructor;
