+++ src/main/java/com/gxdxx/instagram/domain/post/exception/PostNotFoundException.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.exception;
+package com.gxdxx.instagram.domain.post.exception;
 
 public class PostNotFoundException extends RuntimeException {
 }
