+++ src/main/java/com/gxdxx/instagram/domain/reply/api/ReplyController.java
@@ -3,7 +3,7 @@ package com.gxdxx.instagram.domain.reply.api;
 import com.gxdxx.instagram.dto.request.ReplyRegisterRequest;
 import com.gxdxx.instagram.dto.request.ReplyUpdateRequest;
 import com.gxdxx.instagram.dto.response.*;
-import com.gxdxx.instagram.exception.InvalidRequestException;
+import com.gxdxx.instagram.global.error.InvalidRequestException;
 import com.gxdxx.instagram.domain.reply.application.ReplyCreateService;
 import com.gxdxx.instagram.domain.reply.application.ReplyDeleteService;
 import com.gxdxx.instagram.domain.reply.application.ReplyUpdateService;
