+++ src/main/java/com/gxdxx/instagram/domain/reply/application/ReplyCreateService.java
@@ -5,8 +5,8 @@ import com.gxdxx.instagram.dto.response.ReplyRegisterResponse;
 import com.gxdxx.instagram.domain.comment.domain.Comment;
 import com.gxdxx.instagram.domain.reply.domain.Reply;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.exception.CommentNotFoundException;
-import com.gxdxx.instagram.exception.UserNotFoundException;
+import com.gxdxx.instagram.domain.comment.exception.CommentNotFoundException;
+import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.comment.dao.CommentRepository;
 import com.gxdxx.instagram.domain.reply.dao.ReplyRepository;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
