+++ src/main/java/com/gxdxx/instagram/domain/user/application/UserLogoutService.java
@@ -3,8 +3,8 @@ package com.gxdxx.instagram.domain.user.application;
 import com.gxdxx.instagram.config.jwt.TokenProvider;
 import com.gxdxx.instagram.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.exception.RefreshTokenInvalidException;
-import com.gxdxx.instagram.exception.UserNotFoundException;
+import com.gxdxx.instagram.global.auth.RefreshTokenInvalidException;
+import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import com.gxdxx.instagram.global.redis.RedisService;
 import lombok.RequiredArgsConstructor;
