+++ src/main/java/com/gxdxx/instagram/domain/user/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.exception;
+package com.gxdxx.instagram.domain.user.exception;
 
 public class UserNotFoundException extends RuntimeException {
 }
