+++ src/main/java/com/gxdxx/instagram/global/auth/AccessTokenCreateService.java
@@ -4,8 +4,7 @@ import com.gxdxx.instagram.config.jwt.TokenProvider;
 import com.gxdxx.instagram.dto.request.AccessTokenCreateRequest;
 import com.gxdxx.instagram.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.exception.RefreshTokenInvalidException;
-import com.gxdxx.instagram.exception.UserNotFoundException;
+import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import com.gxdxx.instagram.global.redis.RedisService;
 import jakarta.servlet.http.HttpServletResponse;
