+++ src/main/java/com/gxdxx/instagram/global/auth/RefreshTokenInvalidException.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.exception;
+package com.gxdxx.instagram.global.auth;
 
 public class RefreshTokenInvalidException extends RuntimeException {
 }
