+++ src/main/java/com/gxdxx/instagram/global/auth/UnauthorizedAccessException.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.exception;
+package com.gxdxx.instagram.global.auth;
 
 public class UnauthorizedAccessException extends RuntimeException {
 }
