+++ src/main/java/com/gxdxx/instagram/global/error/InvalidRequestException.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.exception;
+package com.gxdxx.instagram.global.error;
 
 public class InvalidRequestException extends RuntimeException {
 }
