+++ src/test/java/com/gxdxx/instagram/service/follow/FollowCreateServiceTest.java
@@ -5,9 +5,9 @@ import com.gxdxx.instagram.dto.request.FollowCreateRequest;
 import com.gxdxx.instagram.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.follow.domain.Follow;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.exception.FollowAlreadyExistsException;
-import com.gxdxx.instagram.exception.InvalidRequestException;
-import com.gxdxx.instagram.exception.UserNotFoundException;
+import com.gxdxx.instagram.domain.follow.exception.FollowAlreadyExistsException;
+import com.gxdxx.instagram.global.error.InvalidRequestException;
+import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.follow.dao.FollowRepository;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import org.junit.jupiter.api.Assertions;
