+++ src/test/java/com/gxdxx/instagram/service/follow/FollowDeleteServiceTest.java
@@ -4,8 +4,8 @@ import com.gxdxx.instagram.domain.follow.application.FollowDeleteService;
 import com.gxdxx.instagram.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.follow.domain.Follow;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.exception.FollowNotFountException;
-import com.gxdxx.instagram.exception.UserNotFoundException;
+import com.gxdxx.instagram.domain.follow.exception.FollowNotFountException;
+import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.follow.dao.FollowRepository;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import org.junit.jupiter.api.Assertions;
