+++ src/test/java/com/gxdxx/instagram/service/user/UserCreateServiceTest.java
@@ -4,7 +4,7 @@ import com.gxdxx.instagram.domain.user.application.UserCreateService;
 import com.gxdxx.instagram.dto.request.UserSignUpRequest;
 import com.gxdxx.instagram.dto.response.UserSignUpResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.exception.NicknameAlreadyExistsException;
+import com.gxdxx.instagram.domain.user.exception.NicknameAlreadyExistsException;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import com.gxdxx.instagram.config.s3.S3Uploader;
 import org.junit.jupiter.api.DisplayName;
