+++ src/test/java/com/gxdxx/instagram/service/user/UserProfileQueryServiceTest.java
@@ -3,7 +3,7 @@ package com.gxdxx.instagram.service.user;
 import com.gxdxx.instagram.domain.user.application.UserProfileQueryService;
 import com.gxdxx.instagram.dto.response.UserProfileResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.exception.UserNotFoundException;
+import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.follow.dao.FollowRepository;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import org.junit.jupiter.api.Assertions;
