+++ src/test/java/com/gxdxx/instagram/service/user/UserProfileUpdateServiceTest.java
@@ -4,8 +4,8 @@ import com.gxdxx.instagram.domain.user.application.UserProfileUpdateService;
 import com.gxdxx.instagram.dto.request.UserProfileUpdateRequest;
 import com.gxdxx.instagram.dto.response.UserProfileUpdateResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.exception.NicknameAlreadyExistsException;
-import com.gxdxx.instagram.exception.UserNotFoundException;
+import com.gxdxx.instagram.domain.user.exception.NicknameAlreadyExistsException;
+import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import com.gxdxx.instagram.config.s3.S3Uploader;
 import org.junit.jupiter.api.Assertions;
